- changes after 2017 - lot of changes might not be tested
Former-commit-id: 57e76f89f5390483280a0dc90ad9ccec6b7d64f5

+++ ggc-desktop-app/ggc-desktop/src/main/java/ggc/gui/dialogs/doctor/DoctorDialog.java
@@ -176,7 +176,7 @@ public class DoctorDialog extends StandardDialogForObject
             this.doctorH.setActiveTill(this.dcTill.getDate());
         }
         this.doctorH.setComment(getText(this.taComment));
-        this.doctorH.setPersonId((int) this.dataAccess.getCurrentUserId());
+        this.doctorH.setPersonId(this.dataAccess.getCurrentUserId());
 
         GGCDb db = ((DataAccess) dataAccess).getDb();
